+++ benchmarks/battleship/app/proguard-rules.pro
