+++ benchmarks/battleship/app/src/main/res/mipmap-xhdpi/ic_launcher.webp
