+++ benchmarks/battleship/app/src/main/res/values/colors.xml
