+++ benchmarks/battleship/app/src/main/res/values/strings.xml
