+++ benchmarks/battleship/benchmark/src/main/AndroidManifest.xml
