--- integration-tests/battleship/src/main/java/com/android/designcompose/testapp/battleship/ui/InterFont.kt
@@ -1,43 +1,0 @@
-/*
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.android.designcompose.testapp.battleship.ui
-
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontStyle
-import androidx.compose.ui.text.font.FontWeight
-import com.android.designcompose.testapp.battleship.R
-
-val interFont =
-    FontFamily(
-        Font(R.font.inter_black, FontWeight.Black),
-        Font(R.font.inter_blackitalic, FontWeight.Black, FontStyle.Italic),
-        Font(R.font.inter_bold, FontWeight.Bold),
-        Font(R.font.inter_bolditalic, FontWeight.Bold, FontStyle.Italic),
-        Font(R.font.inter_extrabold, FontWeight.ExtraBold),
-        Font(R.font.inter_extrabolditalic, FontWeight.ExtraBold, FontStyle.Italic),
-        Font(R.font.inter_extralight, FontWeight.ExtraLight),
-        Font(R.font.inter_extralightitalic, FontWeight.ExtraLight, FontStyle.Italic),
-        Font(R.font.inter_italic, FontWeight.Normal, FontStyle.Italic),
-        Font(R.font.inter_medium, FontWeight.Medium),
-        Font(R.font.inter_mediumitalic, FontWeight.Medium, FontStyle.Italic),
-        Font(R.font.inter_regular, FontWeight.Normal),
-        Font(R.font.inter_semibold, FontWeight.SemiBold),
-        Font(R.font.inter_semibolditalic, FontWeight.SemiBold, FontStyle.Italic),
-        Font(R.font.inter_thin, FontWeight.Thin),
-        Font(R.font.inter_thinitalic, FontWeight.Thin, FontStyle.Italic),
-    )
