--- integration-tests/benchmark-battleship/src/main/java/com/android/designcompose/benchmark/battleship/StartupBenchmarks.kt
@@ -1,94 +1,0 @@
-/*
- * Copyright 2023 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.android.designcompose.benchmark.battleship
-
-import androidx.benchmark.macro.CompilationMode
-import androidx.benchmark.macro.StartupMode
-import androidx.benchmark.macro.StartupTimingMetric
-import androidx.benchmark.macro.junit4.MacrobenchmarkRule
-import androidx.test.ext.junit.runners.AndroidJUnit4
-import org.junit.Rule
-import org.junit.Test
-import org.junit.runner.RunWith
-
-/**
- * This is an example startup benchmark.
- *
- * It navigates to the device's home screen, and launches the default activity.
- *
- * Before running this benchmark:
- * 1) switch your app's active build variant in the Studio (affects Studio runs only)
- * 2) add `<profileable android:shell="true" />` to your app's manifest, within the `<application>`
- *    tag
- *
- * Run this benchmark from Studio to see startup measurements, and captured system traces for
- * investigating your app's performance.
- */
-@RunWith(AndroidJUnit4::class)
-class StartupBenchmarks {
-    @get:Rule val benchmarkRule = MacrobenchmarkRule()
-
-    @Test
-    fun coldStartup() =
-        benchmarkRule.measureRepeated(
-            packageName = "com.android.designcompose.testapp.battleship",
-            metrics = listOf(StartupTimingMetric()),
-            iterations = 5,
-            startupMode = StartupMode.COLD
-        ) {
-            pressHome()
-            startActivityAndWait()
-        }
-
-    @Test
-    fun hotStartup() =
-        benchmarkRule.measureRepeated(
-            packageName = "com.android.designcompose.testapp.battleship",
-            metrics = listOf(StartupTimingMetric()),
-            iterations = 5,
-            startupMode = StartupMode.HOT
-        ) {
-            pressHome()
-            startActivityAndWait()
-        }
-
-    @Test
-    fun noPrecompileStartup() =
-        benchmarkRule.measureRepeated(
-            packageName = "com.android.designcompose.testapp.battleship",
-            metrics = listOf(StartupTimingMetric()),
-            iterations = 5,
-            compilationMode = CompilationMode.None(),
-            startupMode = StartupMode.COLD
-        ) {
-            pressHome()
-            startActivityAndWait()
-        }
-
-    @Test
-    fun fullPrecompileStartup() =
-        benchmarkRule.measureRepeated(
-            packageName = "com.android.designcompose.testapp.battleship",
-            metrics = listOf(StartupTimingMetric()),
-            iterations = 5,
-            compilationMode = CompilationMode.Full(),
-            startupMode = StartupMode.COLD
-        ) {
-            pressHome()
-            startActivityAndWait()
-        }
-}
